requirements.txt
flask==2.3.2
deepseek-sdk==1.0.0